third_party/WebKit/LayoutTests/fast/dom/Range/getClientRects-character-expected.txt

Test horizontal
PASS 1 is rects.length
PASS box.offsetWidth / 26 is >= rects[0].width
PASS box.offsetHeight / 2 is >= rects[0].height
Test vertical
PASS 1 is rects.length
PASS box.offsetWidth / 2 is >= rects[0].width
PASS box.offsetHeight / 26 is >= rects[0].height
ABCDEFGHIJKLMNOPQRSTUVWXYZ ABCDEFGHIJKLMNOPQRSTUVWXYZ
ABCDEFGHIJKLMNOPQRSTUVWXYZ ABCDEFGHIJKLMNOPQRSTUVWXYZ
